// frontend/src/api/generated/models/paymentRequestResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface PaymentRequestResponse {
  id: string;
  request_number: string;
  request_type: string;
  amount: string;
  recipient: string;
  due_date: string | null;
  status: string;
}